Fix Windows scaling issue for real
This might have its roots in DPF. When screen scaling is > 100% and
the host process is DPI-aware, either the hosts passes scaled down
dimensions to DPF or some faulty arithmetic ends up passing scaled down
values to the UI class constructor.

Solution:

1. Query if process (aka the host) is DPI-aware
2. If yes, obtain the current screen scale factor and apply to the
   width/height values passed to UI ctor
3. If no, use the "original" unscaled width/height for ctor

Steps 1 and 2 require winapi functions that MinGW fails to see in the
corresponding header, this is fixed by loading the required functions
from a DLL during runtime.