Disable DSSI
Not compatible with DISTRHO_PLUGIN_WANT_FULL_STATE. Probably it will never come back, build was not working as well. Providing VST2 and LV2 sounds good enough.

Having a target standalone target for Windows sounds interesting and it is currently missing. Having standalone targets on all platforms without depending on JACK sounds even more interesting but probably out of scope for DPF. An acceptable workaround would be bundling the JACK server binary for Mac and Windows. Or something like a "private" JACK that can be statically linked into the standalone binary. If system JACK is running, connect to it. If not, launch an internal server and connect to it. The internal server would not be visible to other processes.

lib/dpf/distrho/src/DistrhoPluginLADSPA+DSSI.cpp:23:3: error: #error Cannot use full state with LADSPA or DSSI
   23 | # error Cannot use full state with LADSPA or DSSI