Change Linux _helper bin suffix to _ui
It is redundant in the d_dpf_webui case because it produces a binary
named d_dpf_webui_ui, but real plugin names are expected not to end in
-ui.

Looks better on ps output, and killing a process named someplugin_ui now
has an expectable consequence. User should not be doing that though.