Remove special case
While for some reason on Linux and Mac it is not needed to set the web
widget size in the UI constructor because UI::uiReshape() is called
later, on Windows omitting the initial size configuration leads to a
blank web widget.

Explicitly setting the web widget size in the UI constructor may be
considered cleaner code so it is not harmless.

Still there is probably some subtle difference between in the Windows DGL.