Make web widgets SubWidget (again)
There cannot be two TopLevelWidget instances at the same time.
Previously ProxyUI (which is a TopLevelWidget) was creating an WebWidget
(another TopLevelWidget) and that leads to keyboard input issues.

To make the web widgets TopLevelWidget the DPF framework needs changes
in DistrhoUI.hpp so the web widget becomes one of the possible UIWidget
implementations, possible the UI_TYPE = html feature. Later...