Rollback 680005f
It is just not possible. Loading manifests from a directory different to
the plugin DLL directory requires "configuration" data and that is not
embeddable, it needs to sit either at the same directory as the DLL,
defeating the original patch purpose, or doing it programmatically which
is possibly too complex for the plugin case.

https://social.msdn.microsoft.com/Forums/vstudio/en-US/b1549de4-87e6-4114-9ff7-5a294d38206b/embed-appconfig-file